+++ src/main/java/com/yanny/ytech/generation/YTechItemTagsProvider.java
@@ -83,6 +83,7 @@ class YTechItemTagsProvider extends ItemTagsProvider {
         tag(YTechItemTags.PRIMITIVE_SMELTERS).add(YTechItems.PRIMITIVE_SMELTER.get());
         tag(YTechItemTags.REINFORCED_BRICKS).add(YTechItems.REINFORCED_BRICKS.get());
         tag(YTechItemTags.REINFORCED_BRICK_CHIMNEYS).add(YTechItems.REINFORCED_BRICK_CHIMNEY.get());
+        tag(YTechItemTags.STRAINERS).add(YTechItems.STRAINER.get());
         tag(YTechItemTags.TERRACOTTA_BRICKS).add(YTechItems.TERRACOTTA_BRICKS.get());
         tag(YTechItemTags.TERRACOTTA_BRICK_SLABS).add(YTechItems.TERRACOTTA_BRICK_SLAB.get());
         tag(YTechItemTags.TERRACOTTA_BRICK_STAIRS).add(YTechItems.TERRACOTTA_BRICK_STAIRS.get());
@@ -151,6 +152,7 @@ class YTechItemTagsProvider extends ItemTagsProvider {
         typedTag(YTechItems.INGOTS, YTechItemTags.INGOTS, MaterialType.VANILLA_METALS);
         typedTag(YTechItems.KNIVES, YTechItemTags.KNIVES);
         typedTag(YTechItems.LEGGINGS, YTechItemTags.LEGGINGS, EnumSet.of(MaterialType.GOLD, MaterialType.IRON, MaterialType.LEATHER));
+        typedTag(YTechItems.MESHES, YTechItemTags.MESHES);
         typedTag(YTechItems.MORTAR_AND_PESTLES, YTechItemTags.MORTAR_AND_PESTLES);
         typedTag(YTechItems.PICKAXES, YTechItemTags.PICKAXES, EnumSet.of(MaterialType.GOLD, MaterialType.IRON));
         typedTag(YTechItems.PLATES, YTechItemTags.PLATES);